requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
black==23.3.0
cachelib==0.10.2
click==8.1.3
Faker==18.6.2
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
inflect==6.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kafka-python==2.0.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.3
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
pycountry==22.3.5
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.9
tenacity==8.2.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.2.3
